API review: move avocado.utils.virt to avocado.core.virt
This module depends closely on the, already moved, avocado.core.remoter
module. Let's move it along.
